rename PipelineExplorerPath, internal vars
