+++ js_modules/dagit/packages/core/src/pipelines/PipelineExplorerRoot.tsx
@@ -5,115 +5,27 @@ import {RouteComponentProps, useHistory} from 'react-router-dom';
 import {useFeatureFlags} from '../app/Flags';
 import {useDocumentTitle} from '../hooks/useDocumentTitle';
 import {Loading} from '../ui/Loading';
-import {NonIdealState} from '../ui/NonIdealState';
 import {buildPipelineSelector} from '../workspace/WorkspaceContext';
 import {AssetGraphExplorer} from '../workspace/asset-graph/AssetGraphExplorer';
 import {RepoAddress} from '../workspace/types';
-import {workspacePathFromAddress} from '../workspace/workspacePath';
 
 import {explodeCompositesInHandleGraph} from './CompositeSupport';
-import {NonIdealPipelineQueryResult} from './NonIdealPipelineQueryResult';
 import {
   GraphExplorer,
   GraphExplorerOptions,
   GRAPH_EXPLORER_FRAGMENT,
   GRAPH_EXPLORER_SOLID_HANDLE_FRAGMENT,
 } from './GraphExplorer';
+import {NonIdealPipelineQueryResult} from './NonIdealPipelineQueryResult';
 import {
-  PipelineExplorerPath,
+  ExplorerPath,
   explorerPathFromString,
   explorerPathToString,
 } from './PipelinePathUtils';
 import {
-  GraphExplorerRootQuery,
-  GraphExplorerRootQueryVariables,
-} from './types/GraphExplorerRootQuery';
-import {
   PipelineExplorerRootQuery,
   PipelineExplorerRootQueryVariables,
 } from './types/PipelineExplorerRootQuery';
-
-export const GraphExplorerRoot: React.FC<RouteComponentProps & {repoAddress: RepoAddress}> = (
-  props,
-) => {
-  const explorerPath = explorerPathFromString(props.match.params['0']);
-  const {repoAddress} = props;
-  const history = useHistory();
-  const [options, setOptions] = React.useState<GraphExplorerOptions>({
-    explodeComposites: false,
-  });
-
-  const selectedName = explorerPath.pathSolids[explorerPath.pathSolids.length - 1];
-  const parentNames = explorerPath.pathSolids.slice(0, explorerPath.pathSolids.length - 1);
-  const pipelineSelector = buildPipelineSelector(repoAddress || null, explorerPath.pipelineName);
-
-  useDocumentTitle(`Graph: ${explorerPath.pipelineName}`);
-
-  const graphResult = useQuery<GraphExplorerRootQuery, GraphExplorerRootQueryVariables>(
-    GRAPH_EXPLORER_ROOT_QUERY,
-    {
-      variables: {
-        graphSelector: {
-          repositoryName: pipelineSelector.repositoryName,
-          repositoryLocationName: pipelineSelector.repositoryLocationName,
-          graphName: explorerPath.pipelineName,
-        },
-        rootHandleID: parentNames.join('.'),
-        requestScopeHandleID: options.explodeComposites ? undefined : parentNames.join('.'),
-      },
-    },
-  );
-
-  return (
-    <Loading<GraphExplorerRootQuery> queryResult={graphResult}>
-      {({graphOrError: result}) => {
-        if (result.__typename === 'GraphNotFoundError') {
-          return (
-            <NonIdealState icon="error" title={'Graph not found'} description={result.message} />
-          );
-        }
-        if (result.__typename === 'PythonError') {
-          return <NonIdealState icon="error" title="Query Error" description={result.message} />;
-        }
-        const parentHandle = result.solidHandle;
-        const displayedHandles = options.explodeComposites
-          ? explodeCompositesInHandleGraph(result.solidHandles)
-          : result.solidHandles;
-
-        const selectedHandle = displayedHandles.find((h) => h.solid.name === selectedName);
-        return (
-          <GraphExplorer
-            options={options}
-            setOptions={setOptions}
-            explorerPath={explorerPath}
-            onChangeExplorerPath={(path, mode) => {
-              const fullPath = workspacePathFromAddress(
-                props.repoAddress,
-                `/graphs/${explorerPathToString(path)}`,
-              );
-              if (mode === 'push') {
-                history.push(fullPath);
-              } else {
-                history.replace(fullPath);
-              }
-            }}
-            pipeline={result}
-            repoAddress={repoAddress}
-            handles={displayedHandles}
-            parentHandle={parentHandle ? parentHandle : undefined}
-            selectedHandle={selectedHandle}
-            isGraph={true}
-            getInvocations={(definitionName) =>
-              displayedHandles
-                .filter((s) => s.solid.definition.name === definitionName)
-                .map((s) => ({handleID: s.handleID}))
-            }
-          />
-        );
-      }}
-    </Loading>
-  );
-};
 
 export const PipelineExplorerSnapshotRoot: React.FC<RouteComponentProps> = (props) => {
   const explorerPath = explorerPathFromString(props.match.params['0']);
@@ -133,8 +45,8 @@ export const PipelineExplorerSnapshotRoot: React.FC<RouteComponentProps> = (prop
 };
 
 export const PipelineExplorerContainer: React.FC<{
-  explorerPath: PipelineExplorerPath;
-  onChangeExplorerPath: (path: PipelineExplorerPath, mode: 'replace' | 'push') => void;
+  explorerPath: ExplorerPath;
+  onChangeExplorerPath: (path: ExplorerPath, mode: 'replace' | 'push') => void;
   repoAddress?: RepoAddress;
   isGraph?: boolean;
 }> = ({explorerPath, repoAddress, onChangeExplorerPath, isGraph = false}) => {
@@ -238,7 +150,7 @@ export const PipelineExplorerContainer: React.FC<{
             setOptions={setOptions}
             explorerPath={explorerPath}
             onChangeExplorerPath={onChangeExplorerPath}
-            pipeline={result}
+            pipelineOrGraph={result}
             repoAddress={repoAddress}
             handles={displayedHandles}
             parentHandle={parentHandle ? parentHandle : undefined}
@@ -305,42 +217,6 @@ export const PIPELINE_EXPLORER_ROOT_QUERY = gql`
       }
       ... on PythonError {
         message
-      }
-    }
-  }
-  ${GRAPH_EXPLORER_FRAGMENT}
-  ${GRAPH_EXPLORER_SOLID_HANDLE_FRAGMENT}
-`;
-
-export const GRAPH_EXPLORER_ROOT_QUERY = gql`
-  query GraphExplorerRootQuery(
-    $graphSelector: GraphSelector
-    $rootHandleID: String!
-    $requestScopeHandleID: String
-  ) {
-    graphOrError(selector: $graphSelector) {
-      ... on Graph {
-        id
-        name
-        ...GraphExplorerFragment
-
-        solidHandle(handleID: $rootHandleID) {
-          ...GraphExplorerSolidHandleFragment
-        }
-        solidHandles(parentHandleID: $requestScopeHandleID) {
-          handleID
-          solid {
-            name
-          }
-          ...GraphExplorerSolidHandleFragment
-        }
-      }
-      ... on GraphNotFoundError {
-        message
-      }
-      ... on PythonError {
-        message
-        stack
       }
     }
   }
